campaign-assistant/requirements.txt
# Phase 1 — local inference
streamlit>=1.35.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.7.0

# Phase 1 — voice input (local GPU transcription + browser mic)
faster-whisper>=1.0.0
audio-recorder-streamlit>=0.0.8

# Phase 2 — cloud LLM providers (uncomment when implementing each client)
# anthropic>=0.25.0
# openai>=1.30.0
# google-generativeai>=0.7.0
# groq>=0.9.0
# mistralai>=1.0.0

# Phase 2 — credential encryption (uncomment alongside credentials.py)
# cryptography>=42.0.0
